backend/requirements.txt
Flask-SQLAlchemy==2.2
Flask-Script==2.0.5
toolz==0.8.2
alembic==0.9.2
psycopg2==2.7.1
pytest==3.1.1
Flask-RESTful==0.3.6
pytest-flask==0.10.0
Flask-Cors==3.0.3
